incorporated db calls with json validator - entries only get written to DB if they pass. added email feature in case any entries fail the validation
