+++ webscraping/node/course-seq-scraper/package.json
@@ -2,12 +2,13 @@
   "name": "concordia-sequence-scrape",
   "version": "1.0.0",
   "description": "Scrape Concordia website to convert course sequences to JSON",
-  "main": "main.js",
+  "main": "scraper.js",
   "author": "Tranzones",
   "dependencies": {
     "cheerio": "latest",
     "mongodb": "^2.2.27",
-    "request": "latest",
-    "remove" : "latest"
+    "nodemailer": "^4.0.1",
+    "remove": "latest",
+    "request": "latest"
   }
 }
