--- webscraping/node/scrapeTheWeb.sh
@@ -1,7 +1,0 @@
-#!/usr/bin/env bash
-
-cd /home/david/CoursePlanner/webscraping/node/course-seq-scraper
-COURSE_PLANNER_HOME="/opt/tomcat/webapps/courses" node driver.js -v
-
-chown tomcat /opt/tomcat/webapps/courses/sequences/
-chgrp tomcat /opt/tomcat/webapps/courses/sequences/
