+++ webscraping/scrapeTheWeb.sh
@@ -1,0 +1,7 @@
+#!/usr/bin/env bash
+
+# run scraper for course sequences
+cd /home/david/CoursePlanner/webscraping/node/course-seq-scraper
+node scrapeAndValidate.js
+
+# TODO: run scraper for course data ...
